优化智能加数、批量加数

+++ electron/service/database/sqlitedb.js
@@ -90,7 +90,7 @@ class SqlitedbService extends Service {
           `CREATE TABLE ball_sum (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         ball_num INTEGER,
-        amount NUMERIC
+        amount INTEGER
       , period INTEGER);`
         this.demoSqliteDB.db.exec(create_table_ball_sum);
       }
